backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
together==0.2.7
python-docx==0.8.11
PyPDF2==3.0.1